criação deposito e saque em conta

+++ src/main/java/br/com/dbserver/bancoapi/controller/dto/ContaDTO.java
@@ -11,7 +11,7 @@ public class ContaDTO {
 
     private Long id;
     private int agencia;
-    private int conta;
+    private Long conta;
     private LocalDate dataCriacaoConta;
     private double saldo;
