+++ src/main/java/br/com/dbserver/bancoapi/controller/dto/NovaContaDTO.java
@@ -11,7 +11,7 @@ public class NovaContaDTO {
 
     private Long id;
     private int agencia;
-    private int conta;
+    private Long conta;
     private double saldo;
     private Cliente cliente;
